+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## add:
 - fpctoolkit.project.rebuild can be called with a shortcut (Shift+F9) - does not need a need to be called from
 - editor commands (fpctoolkit.editor.): base64encode, base64decode, hexencode, hexdecode, urlencode, urldecode
+- ctrl+shift+g Generate UUID
 ## fix:
 - activate project - reloadTasks no longer work (used async call), client reloaded now
 - refactor rename - support for all units